+++ app/src/main/java/br/com/stone/emeraldcomponentsandroid/activities/CounterActivity.kt
@@ -3,6 +3,7 @@ package br.com.stone.emeraldcomponentsandroid.activities
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import br.com.stone.emeraldcomponentsandroid.R
+import kotlinx.android.synthetic.main.activity_counter.*
 
 class CounterActivity : AppCompatActivity() {
 
@@ -10,5 +11,9 @@ class CounterActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_counter)
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
+
+        emeraldCounter.run {
+            setup(0, 3, 1)
+        }
     }
 }